Move PointData template class to its own files and use extern templates
Also:
- increase cmake min version to 3.0
- add include directory in cmake to avoid relative paths in includes
- de-template-ize simple normal computation and move to anonymous namespace

Note: Unfortunately I cannot move other templates out because of the all-or-nothing structure of the Poisson files
